Rename primary to host
The Overwitch library uses primary to Overbridge (p2o) and Overbridge to primary (o2p)
acronyms to specify the data flow direction.
These are later redefined as JACK to Overbridge (j2o) and Overbridge to JACK (o2j) in
the specific client.
Only a JACK client is currently written but it is possible to have new clients in the
future, like a PipeWire one.