Disable optional H2O
The reason for removing this and enabling H2O always is that callbacks
for getting the latency happens before the latency calculation is stable
even though the resampler ratio is because there is no audio flowing.

Since the recommended quality value is 2 (medium quality) and it is also
the default quality value used, it is safe to remove this feature as the
difference in CPU usage is not that relevant.