Add retune state to resampler
The idea is having a state to do exactly the same as with the tune state
with the addition of audio processing. This might create audio artifacts,
specially when the xrun comes with big fluctuations in time measurements,
but avoids a complete resetting of the resampler, which skips the audio
processing for as long as 7 s.